+++ R/doc2vec.R
@@ -10,9 +10,9 @@ as.matrix.textmodel_docvector <- function(x, ...){
 #' @param x a [quanteda::tokens] or [quanteda::dfm] object.
 #' @param model a textmodel_wordvector object.
 #' @param normalize if `TRUE`, normalized word vectors before creating document vectors.
-#' @param weight weight the word vectors by user-provided values; either a single value or 
+#' @param weights weight the word vectors by user-provided values; either a single value or 
 #'    multiple values sorted in the same order as the word vectors.
-#' @param pattern [quanteda::pattern] to select words to apply `weight`. 
+#' @param pattern [quanteda::pattern] to select words to apply `weights`. 
 #' @param group_data if `TRUE`, apply `dfm_group(x)` before creating document vectors.
 #' @param ... additional arguments passed to [quanteda::object2id].
 #' @returns Returns a textmodel_docvector object with the following elements:
@@ -25,7 +25,7 @@ as.matrix.textmodel_docvector <- function(x, ...){
 #'   \item{version}{the version of the wordvector package.}
 #' @export
 textmodel_doc2vec <- function(x, model, normalize = FALSE, 
-                              weight = 1.0, pattern = NULL, 
+                              weights = 1.0, pattern = NULL, 
                               group_data = FALSE, ...) {
     UseMethod("textmodel_doc2vec")
 }
@@ -33,7 +33,7 @@ textmodel_doc2vec <- function(x, model, normalize = FALSE,
 #' @export
 #' @method textmodel_doc2vec tokens
 textmodel_doc2vec.tokens <- function(x, model, normalize = FALSE, 
-                                     weight = 1.0, pattern = NULL, 
+                                     weights = 1.0, pattern = NULL, 
                                      group_data = FALSE, ...) {
     
     if (!identical(class(model), "textmodel_wordvector"))
@@ -41,7 +41,7 @@ textmodel_doc2vec.tokens <- function(x, model, normalize = FALSE,
 
     x <- dfm(x, remove_padding = TRUE, tolower = FALSE)
     result <- textmodel_doc2vec(x, model = model, normalize = normalize,
-                                weight = weight, pattern = pattern, 
+                                weights = weights, pattern = pattern, 
                                 group_data = group_data, ...)
     result$call <- try(match.call(sys.function(-1), call = sys.call(-1)), silent = TRUE)
     return(result)
@@ -50,7 +50,7 @@ textmodel_doc2vec.tokens <- function(x, model, normalize = FALSE,
 #' @export
 #' @method textmodel_doc2vec dfm
 textmodel_doc2vec.dfm <- function(x, model = NULL, normalize = FALSE, 
-                                  weight = 1.0, pattern = NULL,
+                                  weights = 1.0, pattern = NULL,
                                   group_data = FALSE, ...) {
     
     conc <- meta(x, field = "concatenator", type = "object")
@@ -58,18 +58,18 @@ textmodel_doc2vec.dfm <- function(x, model = NULL, normalize = FALSE,
     
     if (is.null(pattern)) {
         n <- nrow(wov)
-        if (length(weight) == 1)
-            weight <- rep(weight, n)
-        weight <- check_double(weight, min = 0, min_len = n, max_len = n)
-        wov <- wov * weight 
+        if (length(weights) == 1)
+            weights <- rep(weights, n)
+        weights <- check_double(weights, min = 0, min_len = n, max_len = n)
+        wov <- wov * weights 
     } else {
-        weight <- check_double(weight, min = 0)
+        weights <- check_double(weights, min = 0)
         ids <- object2id(pattern, types = rownames(wov), match_pattern = "single", 
                          keep_nomatch = FALSE, concatenator = conc,
                          ...)
         if (length(ids) > 0) {
             i <- unique(unlist(ids, use.names = FALSE))
-            wov[i,] <- wov[i,,drop = FALSE] * rep(weight, length(i))
+            wov[i,] <- wov[i,,drop = FALSE] * rep(weights, length(i))
         }
     }